PN-10188 check availability messages in queue

+++ src/main/java/it/pagopa/pn/cucumber/SqsUtils.java
@@ -1,9 +1,8 @@
 package it.pagopa.pn.cucumber;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.SneakyThrows;
-import software.amazon.awssdk.services.eventbridge.model.PutEventsRequest;
+import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.services.eventbridge.model.PutEventsRequestEntry;
 import software.amazon.awssdk.services.sqs.SqsClient;
 import software.amazon.awssdk.services.sqs.model.Message;
@@ -11,7 +10,9 @@ import software.amazon.awssdk.services.sqs.model.ReceiveMessageResponse;
 
 import java.util.Optional;
 
+@Slf4j
 public class SqsUtils {
+
 
     private static final SqsClient sqsClient = SqsClient.create();
     private static final ObjectMapper objectMapper = new ObjectMapper();
@@ -22,23 +23,34 @@ public class SqsUtils {
         String queueUrl = sqsClient.getQueueUrl(builder -> builder.queueName(queueName)).queueUrl();
         ReceiveMessageResponse response = sqsClient.receiveMessage(builder -> builder.queueUrl(queueUrl));
         Optional<Message> optionalMessage;
-        while (response.hasMessages()) {
-            optionalMessage = response.messages().stream()
-                    .filter(message -> {
-                        String messageBody = message.body();
-                        PutEventsRequestEntry putEventsRequestEntry = convertStringToObject(messageBody, PutEventsRequestEntry.class);
-                        NotificationMessage notificationMessage = convertStringToObject(putEventsRequestEntry.detail(), NotificationMessage.class);
-                        return putEventsRequestEntry.detailType().equals(EVENT_BUS_SOURCE_AVAILABLE_DOCUMENT) && putEventsRequestEntry.source().equals(GESTORE_DISPONIBILITA_EVENT_NAME) && notificationMessage.getKey().equals(fileKey);
-                    })
-                    .findFirst();
-            if (optionalMessage.isPresent()) {
-                Message message = optionalMessage.get();
-                sqsClient.deleteMessage(builder -> builder.queueUrl(queueUrl).receiptHandle(message.receiptHandle()));
-                return true;
+        long startTime = System.currentTimeMillis();
+        long endTime = startTime + 15000; // 15 secondi
+
+        while (System.currentTimeMillis() < endTime){
+            if (response.hasMessages()) {
+                optionalMessage = response.messages().stream()
+                        .filter(message -> {
+                            String messageBody = message.body();
+                            PutEventsRequestEntry putEventsRequestEntry = convertStringToObject(messageBody, PutEventsRequestEntry.class);
+                            NotificationMessage notificationMessage = convertStringToObject(putEventsRequestEntry.detail(), NotificationMessage.class);
+                            return putEventsRequestEntry.detailType().equals(EVENT_BUS_SOURCE_AVAILABLE_DOCUMENT)
+                                    && putEventsRequestEntry.source().equals(GESTORE_DISPONIBILITA_EVENT_NAME)
+                                    && notificationMessage.getKey().equals(fileKey);
+                        })
+                        .findFirst();
+                if (optionalMessage.isPresent()) {
+                    Message message = optionalMessage.get();
+                    log.debug("message {}", message);
+                    sqsClient.deleteMessage(builder -> builder.queueUrl(queueUrl).receiptHandle(message.receiptHandle()));
+                    return true;
+                }
             }
         }
+
         return false;
     }
+
+
 
     @SneakyThrows
     private static <T> T convertStringToObject(String payload, Class<T> classToConvert) {
